requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
openai==1.14.0
python-dotenv==1.0.0
google-generativeai==0.3.2
anthropic==0.21.3